requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
rev-ai==2.21.0
gunicorn==21.2.0
Werkzeug==2.3.7
pyairtable==1.5.0 